+++ customer/src/components/Category.jsx
@@ -1,27 +1,4 @@
-import styled from 'styled-components'
-import mobile from './Responsive'
-
-const Wrapper = styled.div`
-  margin-bottom: 5px;
-`
-const Button = styled.button`
-  border: none;
-  border-radius:40%
-  width: 45px;
-  height: 20px;
-  left: 47px;
-  top: 151px;
-
-  font-family: Actor;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 17px;
-  line-height: 20px;
-
-  color: #fa4a0c;
-  backgound-color: white;
-  ${mobile({ width: '80px', height: '40px', margin: '3px' })}
-`
+import { Button, Wrapper } from '../styles/category.styles'
 
 const Category = () => {
   return (
