+++ customer/src/styles/Responsive.js
@@ -1,10 +1,10 @@
-import { css } from "styled-components";
+import { css } from 'styled-components'
 
 export const mobile = (props) => {
   return css`
     @media only screen and (max-width: 480px) {
       ${props}
     }
-  `;
-};
-export default mobile;
+  `
+}
+export default mobile
